Simplificados testes e adicionados testes de Produto

+++ src/test/java/bootcamp/mercado/endpointfake/NotaFiscalRequestTest.java
@@ -133,7 +133,7 @@ class NotaFiscalRequestTest {
         List<Produto> produtos = produtoRepository.saveAll(List.of(
                 produtoBuilder
                         .setNome("Produto 1")
-                        .setPreco(BigDecimal.valueOf(20.14))
+                        .setPreco(20.14)
                         .setQuantidade(5)
                         .addCaracteristica(pickRandomValue(caracteristicaRequests))
                         .setCategoria(pickRandomValue(categorias).getNome())
@@ -141,7 +141,7 @@ class NotaFiscalRequestTest {
 
                 produtoBuilder
                         .setNome("Produto 2")
-                        .setPreco(BigDecimal.valueOf(35.20))
+                        .setPreco(35.20)
                         .setQuantidade(1)
                         .addCaracteristica(pickRandomValue(caracteristicaRequests))
                         .setCategoria(pickRandomValue(categorias).getNome())
@@ -149,7 +149,7 @@ class NotaFiscalRequestTest {
 
                 produtoBuilder
                         .setNome("Produto 3")
-                        .setPreco(BigDecimal.valueOf(55.10))
+                        .setPreco(55.10)
                         .setQuantidade(2)
                         .addCaracteristica(pickRandomValue(caracteristicaRequests))
                         .setCategoria(pickRandomValue(categorias).getNome())
