+++ src/test/java/bootcamp/mercado/produto/compra/gateway/response/PaypalRequestTest.java
@@ -127,7 +127,7 @@ class PaypalRequestTest {
         List<Produto> produtos = produtoRepository.saveAll(List.of(
                 produtoBuilder
                         .setNome("Produto 1")
-                        .setPreco(BigDecimal.valueOf(20.14))
+                        .setPreco(20.14)
                         .setQuantidade(5)
                         .addCaracteristica(pickRandomValue(caracteristicaRequests))
                         .setCategoria(pickRandomValue(categorias).getNome())
@@ -135,7 +135,7 @@ class PaypalRequestTest {
 
                 produtoBuilder
                         .setNome("Produto 2")
-                        .setPreco(BigDecimal.valueOf(35.20))
+                        .setPreco(35.20)
                         .setQuantidade(2)
                         .addCaracteristica(pickRandomValue(caracteristicaRequests))
                         .setCategoria(pickRandomValue(categorias).getNome())
@@ -143,7 +143,7 @@ class PaypalRequestTest {
 
                 produtoBuilder
                         .setNome("Produto 3")
-                        .setPreco(BigDecimal.valueOf(55.10))
+                        .setPreco(55.10)
                         .setQuantidade(2)
                         .addCaracteristica(pickRandomValue(caracteristicaRequests))
                         .setCategoria(pickRandomValue(categorias).getNome())
